announceNodeToStream test
