+++ packages/broker/test/integration/plugins/operator/announceNodeToStream.test.ts
@@ -1,0 +1,34 @@
+import { toStreamID } from '@streamr/protocol'
+import { fastPrivateKey } from '@streamr/test-utils'
+import { collect } from '@streamr/utils'
+import { announceNodeToStream } from '../../../../src/plugins/operator/announceNodeToStream'
+import { createClient } from '../../../utils'
+import { setupOperatorContract } from './contractUtils'
+
+const TIMEOUT = 10 * 1000
+
+describe('announceNodeToStream', () => {
+
+    it('publishes to stream', async () => {
+        const { operatorServiceConfig, nodeWallets } = await setupOperatorContract({
+            nodeCount: 1
+        })
+        const operatorContractAddress = operatorServiceConfig.operatorContractAddress
+        const nodeWallet = nodeWallets[0]
+        const client = createClient(nodeWallet.privateKey)
+        const streamId = toStreamID('/operator/coordination', operatorContractAddress)
+        const anonymousClient = createClient(fastPrivateKey())
+        const subscription = await anonymousClient.subscribe(streamId)
+
+        await announceNodeToStream(operatorContractAddress, client)
+
+        const [{ content }] = await collect(subscription, 1)
+        expect(content).toEqual({
+            msgType: 'heartbeat',
+            peerDescriptor: await client.getPeerDescriptor()
+        })
+
+        await anonymousClient.destroy()
+        await client.destroy()
+    }, TIMEOUT)
+})
